consolidate scraper code
