+++ dspback/routers/hydroshare.py
@@ -11,7 +11,7 @@ from dspback.dependencies import RepositoryException
 from dspback.pydantic_schemas import RepositoryType, SubmissionBase
 from dspback.routers.metadata_class import MetadataRoutes
 from dspback.schemas.hydroshare.model import ResourceMetadata
-from dspback.utils.jsonld.hydroshare import scrape_jsonld
+from dspback.utils.jsonld.scraper import scrape_jsonld
 from dspback.utils.mongo import delete_jsonld, upsert_jsonld
 
 router = InferringRouter()
@@ -150,7 +150,7 @@ class HydroShareMetadataRoutes(MetadataRoutes):
         if response.status_code >= 300:
             raise RepositoryException(status_code=response.status_code, detail=response.text)
 
-        json_ld = scrape_jsonld(response.text, identifier)
+        json_ld = scrape_jsonld(response.text, identifier, {"id": "schemaorg"})
         upsert_jsonld(json_ld)
         return json_ld
