--- dspback/utils/jsonld/earthchem.py
@@ -1,15 +1,0 @@
-import json
-
-from bs4 import BeautifulSoup
-
-from dspback.utils.jsonld.formatter import format_fields
-
-
-def scrape_jsonld(resource_data, identifier):
-    resource_soup = BeautifulSoup(resource_data, "html.parser")
-    resource_json_ld = resource_soup.find("script", {"type": "application/ld+json"})
-    resource_json_ld = json.loads(resource_json_ld.text)
-    resource_json_ld = format_fields(resource_json_ld)
-    resource_json_ld["@repository_identifier"] = identifier
-
-    return resource_json_ld
